requirements.txt
streamlit
python-dotenv
PyMuPDF
python-docx
requests
rake-nltk
crewai
pydantic
spacy
